Changed maven build to be more best practice
Standard Directory Layout
https://maven.apache.org/guides/introduction/introduction-to-the-standard-directory-layout.html

Renamed definitions to agrirouter-api-protobuf-definitions
We have the convention in our company that the maven artifacts should have the same name like the repository. I don't see a reason why this shouldn't be done here.

Removed the profile
We are using maven as well for our builds and eclipse. In our default configuration we are using no profiles at all. I don't see a point in using a profile here. You could just delete the documentation in the build configs.

Use java 8
This is not really required but maybe later the google protobuf generators will use Java 8 features and this will make sure that it will work with java 8

Removed source src/main/generated
In maven the default behavior of putting generated sources is to place it in target/generated-sources (https://maven.apache.org/guides/mini/guide-generating-sources.html). We are typically using these projects in eclipse with m2e integration and when using the default folders they are automatically added to the build path in eclipse.