Give proto files unique names
When generating protobuf code for C#, the generator will place all
source files in the same folder, thus overwriting existing source files
when .proto files have the same name.

This is the case for endpoints.proto in the account payload. These
files shall have unique names, like it is done for the feed payload.